Restructured Project: Moved FromAnswer, NES, and base Summarizer -> summarizers
