+++ KBQA/appB/summarizers/base_summarizer/__init__.py
@@ -1,0 +1,1 @@
+"""Root of base summarizer."""
