+++ KBQA/appB/summarizers/from_answer_summarizer/__init__.py
@@ -1,0 +1,1 @@
+"""Root of the FromAnswer Summarizer."""
